Add lsp support to Lua config in neovim
Keybinding mappings for completions do not work in Lua, only
Vimscript seems to work for those, so autocmplete confirmation
from nvim-compe seems to be a no-go for now.

Otherwise the autocomplete features seem to work fine except for
confirmation and exiting of the autocomplete features which lack
true keybindings to make them usable.