Kakoune barebones configuration
Kaoune is really cool. But it lack some proper QOL features to
really make it worth having without doubting.

My biggest gripe was not being able to have an 'expandtab' option
as in neovim for not using tabs. I *should* be easy, it's not.
There's a whole plugin for that. One that has been abandoned by its
author, in favor of emacs.